test: move backward update tests next to the FP8 checkpoint tests in test_numerics.py
